Players pool

+++ ScrabbleJunior/ScrabbleJunior.cpp
@@ -3,11 +3,11 @@
 
 #include "Pool.h"
 #include "Player.h"
-#include "PlayBoard.h"
+#include "Board.h"
 
 int main()
 {
-    srand(time(NULL)); //letter randomize
+    srand((unsigned int) time(NULL)); //letter randomize
 
     //english alphabet 
     std::string englishAlphabet = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
